backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
--no-binary psycopg2

aniso8601==9.0.1
    # via graphene
asgiref==3.7.2
    # via django
beautifulsoup4==4.12.2
    # via django-livereload-server
certifi==2025.10.5
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==46.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth==6.0.0
    # via -r requirements.in
django==4.2.26
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-filter
    #   django-js-asset
    #   django-livereload-server
    #   django-prose-editor
    #   django-revproxy
    #   djangorestframework
    #   graphene-django
django-allauth==0.61.1
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-js-asset==3.1.2
    # via django-prose-editor
django-livereload-server==0.4
    # via -r requirements.in
django-prose-editor[sanitize]==0.15.2
    # via -r requirements.in
django-revproxy==0.12.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
execnet==2.0.2
    # via pytest-xdist
faker==23.2.0
    # via -r requirements.in
graphene==3.3
    # via
    #   -r requirements.in
    #   graphene-django
graphene-django==3.2.2
    # via -r requirements.in
graphql-core==3.2.3
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
nh3==0.2.21
    # via django-prose-editor
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via pytest
pillow==11.3.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
promise==2.3
    # via graphene-django
psycopg2==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
pytest==7.4.3
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-xdist
pytest-django==4.7.0
    # via -r requirements.in
pytest-xdist==3.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python3-openid==3.2.0
    # via django-allauth
requests==2.32.5
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
six==1.16.0
    # via
    #   django-livereload-server
    #   promise
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
text-unidecode==1.3
    # via graphene-django
tornado==6.5.2
    # via django-livereload-server
urllib3==2.5.0
    # via
    #   django-revproxy
    #   requests
